move lua sources to builtin/, add deprecation warnings
Changes

- merge and relocate `lua/compe_buffer/*.lua` files to `lua/compe/builtin/buffer.lua`
- relocate `compe_nvim_lsp/source.lua` has been to `lua/compe/builtin/nvim_lsp.lua`
- add `use` method to let users easily register and use builtin sources

```
require'compe.builtin'.use({"vsnip", "path", "nvim_lsp"})
```

- add deprecation warnings for setting up buffer and nvim_lsp source.
- add inline documentation (not complete yet)
